Fix Type-Error on start

+++ src/ui/run_ui.py
@@ -39,8 +39,8 @@ class Main_Dialog(QMainWindow, Main_UI):
     def center_screen(self):
         size = self.size()
         desktopSize = QtWidgets.QDesktopWidget().screenGeometry()
-        top = (desktopSize.height() / 2) - (size.height() / 2)
-        left = (desktopSize.width() / 2) - (size.width() / 2)
+        top = int((desktopSize.height() / 2) - (size.height() / 2))
+        left = int((desktopSize.width() / 2) - (size.width() / 2))
         self.move(left, top)
 
     def closeEvent(self, event):
